refactor: change api name
