+++ src/components/Search/Search.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import SearchInput from './SearchInput'
 import SearchResult from './SearchResult'
 import { selectSearch } from './searchSlice'
-import { SickObj } from '../../apis/sick'
+import { SickObj } from '../../apis/suggestion'
 import useKeyboardNavigation from '../../hooks/useKeyboardNavigation'
 import useSuggestion from '../../hooks/useSuggestion'
 import { useAppSelector } from '../../store/reduxHooks'
