+++ src/hooks/useSuggestion.ts
@@ -2,7 +2,7 @@ import { useState } from 'react'
 
 import { useCacheContext } from './useCacheContext'
 import { api } from '../apis'
-import { GetSickResponse } from '../apis/sick'
+import { GetSickResponse } from '../apis/suggestion'
 
 function useSuggestion(queryKey: string) {
   const [data, setData] = useState<GetSickResponse>([])
@@ -18,7 +18,7 @@ function useSuggestion(queryKey: string) {
     if (cacheClient.get(searchTerm)) return
     setIsLoading(true)
     try {
-      const res = await api.sick.get(searchTerm)
+      const res = await api.suggestion.get(searchTerm)
       setData(res.data)
       cacheClient.set(searchTerm, res.data)
     } catch (e) {
